+++ go/pipeline/base_batch.go
@@ -6,6 +6,7 @@ package pipeline
 import "C"
 
 import (
+	"errors"
 	"fmt"
 	"math"
 	"reflect"
@@ -17,7 +18,6 @@ import (
 	"github.com/valkey-io/valkey-glide/go/v2/options"
 
 	"github.com/valkey-io/valkey-glide/go/v2/internal"
-	"github.com/valkey-io/valkey-glide/go/v2/internal/errors"
 	"github.com/valkey-io/valkey-glide/go/v2/internal/utils"
 )
 
@@ -1745,7 +1745,7 @@ func (b *BaseBatch[T]) LMPop(keys []string, listDirection constants.ListDirectio
 
 	// Check for potential length overflow.
 	if len(keys) > math.MaxInt-2 {
-		return b.addError("LMPop", &errors.RequestError{Msg: "Length overflow for the provided keys"})
+		return b.addError("LMPop", errors.New("length overflow for the provided keys"))
 	}
 
 	// args slice will have 2 more arguments with the keys provided.
@@ -1784,7 +1784,7 @@ func (b *BaseBatch[T]) LMPopCount(keys []string, listDirection constants.ListDir
 
 	// Check for potential length overflow.
 	if len(keys) > math.MaxInt-4 {
-		return b.addError("LMPopCount", &errors.RequestError{Msg: "Length overflow for the provided keys"})
+		return b.addError("LMPopCount", errors.New("length overflow for the provided keys"))
 	}
 
 	// args slice will have 4 more arguments with the keys provided.
@@ -1829,7 +1829,7 @@ func (b *BaseBatch[T]) BLMPop(keys []string, listDirection constants.ListDirecti
 
 	// Check for potential length overflow.
 	if len(keys) > math.MaxInt-3 {
-		return b.addError("BLMPop", &errors.RequestError{Msg: "Length overflow for the provided keys"})
+		return b.addError("BLMPop", errors.New("length overflow for the provided keys"))
 	}
 
 	// args slice will have 3 more arguments with the keys provided.
@@ -1880,7 +1880,7 @@ func (b *BaseBatch[T]) BLMPopCount(
 
 	// Check for potential length overflow.
 	if len(keys) > math.MaxInt-5 {
-		return b.addError("BLMPopCount", &errors.RequestError{Msg: "Length overflow for the provided keys"})
+		return b.addError("BLMPopCount", errors.New("length overflow for the provided keys"))
 	}
 
 	// args slice will have 5 more arguments with the keys provided.
@@ -2036,7 +2036,7 @@ func (b *BaseBatch[T]) Exists(keys []string) *T {
 // Sets a timeout on key. After the timeout has expired, the key will automatically be deleted.
 //
 // If key already has an existing expire set, the time to live is updated to the new value.
-// If seconds is a non-positive number, the key will be deleted rather than expired.
+// If expireTime is a non-positive number, the key will be deleted rather than expired.
 // The timeout will only be cleared by commands that delete or overwrite the contents of key.
 //
 // See [valkey.io] for details.
@@ -2044,7 +2044,7 @@ func (b *BaseBatch[T]) Exists(keys []string) *T {
 // Parameters:
 //
 //	key - The key to expire.
-//	seconds - Time in seconds for the key to expire
+//	expireTime - Duration for the key to expire
 //
 // Command Response:
 //
@@ -2052,14 +2052,14 @@ func (b *BaseBatch[T]) Exists(keys []string) *T {
 //	or operation skipped due to the provided arguments.
 //
 // [valkey.io]: https://valkey.io/commands/expire/
-func (b *BaseBatch[T]) Expire(key string, seconds int64) *T {
-	return b.addCmdAndTypeChecker(C.Expire, []string{key, utils.IntToString(seconds)}, reflect.Bool, false)
+func (b *BaseBatch[T]) Expire(key string, expireTime time.Duration) *T {
+	return b.addCmdAndTypeChecker(C.Expire, []string{key, utils.FloatToString(expireTime.Seconds())}, reflect.Bool, false)
 }
 
 // Sets a timeout on key. After the timeout has expired, the key will automatically be deleted.
 //
 // If key already has an existing expire set, the time to live is updated to the new value.
-// If seconds is a non-positive number, the key will be deleted rather than expired.
+// If expireTime is a non-positive number, the key will be deleted rather than expired.
 // The timeout will only be cleared by commands that delete or overwrite the contents of key.
 //
 // See [valkey.io] for details.
@@ -2067,7 +2067,7 @@ func (b *BaseBatch[T]) Expire(key string, seconds int64) *T {
 // Parameters:
 //
 //	key - The key to expire.
-//	seconds - Time in seconds for the key to expire.
+//	expireTime - Duration for the key to expire.
 //	expireCondition - The option to set expiry, see [constants.ExpireCondition].
 //
 // Command Response:
@@ -2076,12 +2076,17 @@ func (b *BaseBatch[T]) Expire(key string, seconds int64) *T {
 //	or operation skipped due to the provided arguments.
 //
 // [valkey.io]: https://valkey.io/commands/expire/
-func (b *BaseBatch[T]) ExpireWithOptions(key string, seconds int64, expireCondition constants.ExpireCondition) *T {
+func (b *BaseBatch[T]) ExpireWithOptions(key string, expireTime time.Duration, expireCondition constants.ExpireCondition) *T {
 	expireConditionStr, err := expireCondition.ToString()
 	if err != nil {
 		return b.addError("ExpireWithOptions", err)
 	}
-	return b.addCmdAndTypeChecker(C.Expire, []string{key, utils.IntToString(seconds), expireConditionStr}, reflect.Bool, false)
+	return b.addCmdAndTypeChecker(
+		C.Expire,
+		[]string{key, utils.FloatToString(expireTime.Seconds()), expireConditionStr},
+		reflect.Bool,
+		false,
+	)
 }
 
 // Sets a timeout on key using an absolute Unix timestamp. It takes an absolute Unix timestamp (seconds since January 1, 1970)
@@ -2090,7 +2095,7 @@ func (b *BaseBatch[T]) ExpireWithOptions(key string, seconds int64, expireCondit
 // If key already has an existing expire set, the time to live is updated to the new value.
 // The timeout will only be cleared by commands that delete or overwrite the contents of key
 // If key already has an existing expire set, the time to live is updated to the new value.
-// If seconds is a non-positive number, the key will be deleted rather than expired.
+// If expireTime is a non-positive number, the key will be deleted rather than expired.
 // The timeout will only be cleared by commands that delete or overwrite the contents of key.
 //
 // See [valkey.io] for details.
@@ -2098,7 +2103,7 @@ func (b *BaseBatch[T]) ExpireWithOptions(key string, seconds int64, expireCondit
 // Parameters:
 //
 //	key - The key to expire.
-//	unixTimestampInSeconds - Absolute Unix timestamp
+//	expireTime - The timestamp for expiry.
 //
 // Command Response:
 //
@@ -2106,8 +2111,8 @@ func (b *BaseBatch[T]) ExpireWithOptions(key string, seconds int64, expireCondit
 //	or operation skipped due to the provided arguments.
 //
 // [valkey.io]: https://valkey.io/commands/expireat/
-func (b *BaseBatch[T]) ExpireAt(key string, unixTimestampInSeconds int64) *T {
-	return b.addCmdAndTypeChecker(C.ExpireAt, []string{key, utils.IntToString(unixTimestampInSeconds)}, reflect.Bool, false)
+func (b *BaseBatch[T]) ExpireAt(key string, expireTime time.Time) *T {
+	return b.addCmdAndTypeChecker(C.ExpireAt, []string{key, utils.IntToString(expireTime.Unix())}, reflect.Bool, false)
 }
 
 // Sets a timeout on key using an absolute Unix timestamp. It takes an absolute Unix timestamp (seconds since January 1, 1970)
@@ -2116,7 +2121,7 @@ func (b *BaseBatch[T]) ExpireAt(key string, unixTimestampInSeconds int64) *T {
 // If key already has an existing expire set, the time to live is updated to the new value.
 // The timeout will only be cleared by commands that delete or overwrite the contents of key
 // If key already has an existing expire set, the time to live is updated to the new value.
-// If seconds is a non-positive number, the key will be deleted rather than expired.
+// If expireTime is a non-positive number, the key will be deleted rather than expired.
 // The timeout will only be cleared by commands that delete or overwrite the contents of key.
 //
 // See [valkey.io] for details.
@@ -2124,7 +2129,7 @@ func (b *BaseBatch[T]) ExpireAt(key string, unixTimestampInSeconds int64) *T {
 // Parameters:
 //
 //	key - The key to expire.
-//	unixTimestampInSeconds - Absolute Unix timestamp.
+//	expireTime - The timestamp for expiry.
 //	expireCondition - The option to set expiry - see [constants.ExpireCondition].
 //
 // Command Response:
@@ -2135,7 +2140,7 @@ func (b *BaseBatch[T]) ExpireAt(key string, unixTimestampInSeconds int64) *T {
 // [valkey.io]: https://valkey.io/commands/expireat/
 func (b *BaseBatch[T]) ExpireAtWithOptions(
 	key string,
-	unixTimestampInSeconds int64,
+	expireTime time.Time,
 	expireCondition constants.ExpireCondition,
 ) *T {
 	expireConditionStr, err := expireCondition.ToString()
@@ -2144,7 +2149,7 @@ func (b *BaseBatch[T]) ExpireAtWithOptions(
 	}
 	return b.addCmdAndTypeChecker(
 		C.ExpireAt,
-		[]string{key, utils.IntToString(unixTimestampInSeconds), expireConditionStr},
+		[]string{key, utils.IntToString(expireTime.Unix()), expireConditionStr},
 		reflect.Bool,
 		false,
 	)
@@ -2152,7 +2157,7 @@ func (b *BaseBatch[T]) ExpireAtWithOptions(
 
 // Sets a timeout on key in milliseconds. After the timeout has expired, the key will automatically be deleted.
 // If key already has an existing expire set, the time to live is updated to the new value.
-// If milliseconds is a non-positive number, the key will be deleted rather than expired.
+// If expireTime is a non-positive number, the key will be deleted rather than expired.
 // The timeout will only be cleared by commands that delete or overwrite the contents of key.
 //
 // See [valkey.io] for details.
@@ -2160,7 +2165,7 @@ func (b *BaseBatch[T]) ExpireAtWithOptions(
 // Parameters:
 //
 //	key - The key to set timeout on it.
-//	milliseconds - The timeout in milliseconds.
+//	expireTime - Duration for the key to expire.
 //
 // Command Response:
 //
@@ -2168,13 +2173,13 @@ func (b *BaseBatch[T]) ExpireAtWithOptions(
 //	or operation skipped due to the provided arguments.
 //
 // [valkey.io]: https://valkey.io/commands/pexpire/
-func (b *BaseBatch[T]) PExpire(key string, milliseconds int64) *T {
-	return b.addCmdAndTypeChecker(C.PExpire, []string{key, utils.IntToString(milliseconds)}, reflect.Bool, false)
+func (b *BaseBatch[T]) PExpire(key string, expireTime time.Duration) *T {
+	return b.addCmdAndTypeChecker(C.PExpire, []string{key, utils.IntToString(expireTime.Milliseconds())}, reflect.Bool, false)
 }
 
 // Sets a timeout on key in milliseconds. After the timeout has expired, the key will automatically be deleted.
 // If key already has an existing expire set, the time to live is updated to the new value.
-// If milliseconds is a non-positive number, the key will be deleted rather than expired.
+// If expireTime is a non-positive number, the key will be deleted rather than expired.
 // The timeout will only be cleared by commands that delete or overwrite the contents of key.
 //
 // See [valkey.io] for details.
@@ -2182,7 +2187,7 @@ func (b *BaseBatch[T]) PExpire(key string, milliseconds int64) *T {
 // Parameters:
 //
 //	key - The key to set timeout on it.
-//	milliseconds - The timeout in milliseconds.
+//	expireTime - Duration for the key to expire.
 //	expireCondition - The option to set expiry, see [constants.ExpireCondition].
 //
 // Command Response:
@@ -2191,14 +2196,14 @@ func (b *BaseBatch[T]) PExpire(key string, milliseconds int64) *T {
 //	or operation skipped due to the provided arguments.
 //
 // [valkey.io]: https://valkey.io/commands/pexpire/
-func (b *BaseBatch[T]) PExpireWithOptions(key string, milliseconds int64, expireCondition constants.ExpireCondition) *T {
+func (b *BaseBatch[T]) PExpireWithOptions(key string, expireTime time.Duration, expireCondition constants.ExpireCondition) *T {
 	expireConditionStr, err := expireCondition.ToString()
 	if err != nil {
 		return b.addError("PExpireWithOptions", err)
 	}
 	return b.addCmdAndTypeChecker(
 		C.PExpire,
-		[]string{key, utils.IntToString(milliseconds), expireConditionStr},
+		[]string{key, utils.IntToString(expireTime.Milliseconds()), expireConditionStr},
 		reflect.Bool,
 		false,
 	)
@@ -2215,7 +2220,7 @@ func (b *BaseBatch[T]) PExpireWithOptions(key string, milliseconds int64, expire
 // Parameters:
 //
 //	key - The key to set timeout on it.
-//	unixTimestampInMilliSeconds - The timeout in an absolute Unix timestamp.
+//	expireTime - The timestamp for expiry.
 //
 // Command Response:
 //
@@ -2223,10 +2228,10 @@ func (b *BaseBatch[T]) PExpireWithOptions(key string, milliseconds int64, expire
 //	or operation skipped due to the provided arguments.
 //
 // [valkey.io]: https://valkey.io/commands/pexpireat/
-func (b *BaseBatch[T]) PExpireAt(key string, unixTimestampInMilliSeconds int64) *T {
+func (b *BaseBatch[T]) PExpireAt(key string, expireTime time.Time) *T {
 	return b.addCmdAndTypeChecker(
 		C.PExpireAt,
-		[]string{key, utils.IntToString(unixTimestampInMilliSeconds)},
+		[]string{key, utils.IntToString(expireTime.UnixMilli())},
 		reflect.Bool,
 		false,
 	)
@@ -2243,7 +2248,7 @@ func (b *BaseBatch[T]) PExpireAt(key string, unixTimestampInMilliSeconds int64) 
 // Parameters:
 //
 //	key - The key to set timeout on it.
-//	unixTimestampInMilliSeconds - The timeout in an absolute Unix timestamp.
+//	expireTime - The timestamp for expiry.
 //	expireCondition - The option to set expiry, see [constants.ExpireCondition].
 //
 // Command Response:
@@ -2254,7 +2259,7 @@ func (b *BaseBatch[T]) PExpireAt(key string, unixTimestampInMilliSeconds int64) 
 // [valkey.io]: https://valkey.io/commands/pexpireat/
 func (b *BaseBatch[T]) PExpireAtWithOptions(
 	key string,
-	unixTimestampInMilliSeconds int64,
+	expireTime time.Time,
 	expireCondition constants.ExpireCondition,
 ) *T {
 	expireConditionStr, err := expireCondition.ToString()
@@ -2263,7 +2268,7 @@ func (b *BaseBatch[T]) PExpireAtWithOptions(
 	}
 	return b.addCmdAndTypeChecker(
 		C.PExpireAt,
-		[]string{key, utils.IntToString(unixTimestampInMilliSeconds), expireConditionStr},
+		[]string{key, utils.IntToString(expireTime.UnixMilli()), expireConditionStr},
 		reflect.Bool,
 		false,
 	)
@@ -2969,9 +2974,7 @@ func (b *BaseBatch[T]) BZMPop(keys []string, scoreFilter constants.ScoreFilter, 
 
 	// Check for potential length overflow.
 	if len(keys) > math.MaxInt-3 {
-		return b.addError("BZMPop", &errors.RequestError{
-			Msg: "Length overflow for the provided keys",
-		})
+		return b.addError("BZMPop", errors.New("length overflow for the provided keys"))
 	}
 
 	// args slice will have 3 more arguments with the keys provided.
@@ -3026,9 +3029,7 @@ func (b *BaseBatch[T]) BZMPopWithOptions(
 
 	// Check for potential length overflow.
 	if len(keys) > math.MaxInt-5 {
-		return b.addError("BZMPopWithOptions", &errors.RequestError{
-			Msg: "Length overflow for the provided keys",
-		})
+		return b.addError("BZMPopWithOptions", errors.New("length overflow for the provided keys"))
 	}
 
 	// args slice will have 5 more arguments with the keys provided.
@@ -3330,7 +3331,7 @@ func (b *BaseBatch[T]) XLen(key string) *T {
 //	    These IDs are deleted from the Pending Entries List.
 //
 // [valkey.io]: https://valkey.io/commands/xautoclaim/
-func (b *BaseBatch[T]) XAutoClaim(key string, group string, consumer string, minIdleTime int64, start string) *T {
+func (b *BaseBatch[T]) XAutoClaim(key string, group string, consumer string, minIdleTime time.Duration, start string) *T {
 	return b.XAutoClaimWithOptions(key, group, consumer, minIdleTime, start, *options.NewXAutoClaimOptions())
 }
 
@@ -3367,11 +3368,11 @@ func (b *BaseBatch[T]) XAutoClaimWithOptions(
 	key string,
 	group string,
 	consumer string,
-	minIdleTime int64,
+	minIdleTime time.Duration,
 	start string,
 	options options.XAutoClaimOptions,
 ) *T {
-	args := []string{key, group, consumer, utils.IntToString(minIdleTime), start}
+	args := []string{key, group, consumer, utils.IntToString(minIdleTime.Milliseconds()), start}
 	optArgs, err := options.ToArgs()
 	if err != nil {
 		return b.addError("XAutoClaimWithOptions", err)
@@ -3408,7 +3409,13 @@ func (b *BaseBatch[T]) XAutoClaimWithOptions(
 //	    These IDs are deleted from the Pending Entries List.
 //
 // [valkey.io]: https://valkey.io/commands/xautoclaim/
-func (b *BaseBatch[T]) XAutoClaimJustId(key string, group string, consumer string, minIdleTime int64, start string) *T {
+func (b *BaseBatch[T]) XAutoClaimJustId(
+	key string,
+	group string,
+	consumer string,
+	minIdleTime time.Duration,
+	start string,
+) *T {
 	return b.XAutoClaimJustIdWithOptions(key, group, consumer, minIdleTime, start, *options.NewXAutoClaimOptions())
 }
 
@@ -3445,11 +3452,11 @@ func (b *BaseBatch[T]) XAutoClaimJustIdWithOptions(
 	key string,
 	group string,
 	consumer string,
-	minIdleTime int64,
+	minIdleTime time.Duration,
 	start string,
 	options options.XAutoClaimOptions,
 ) *T {
-	args := []string{key, group, consumer, utils.IntToString(minIdleTime), start}
+	args := []string{key, group, consumer, utils.IntToString(minIdleTime.Milliseconds()), start}
 	optArgs, err := options.ToArgs()
 	if err != nil {
 		return b.addError("XAutoClaimJustIdWithOptions", err)
@@ -3640,7 +3647,7 @@ func (b *BaseBatch[T]) XGroupCreateWithOptions(key string, group string, id stri
 // Parameters:
 //
 //	key - The key to create.
-//	ttl - The expiry time (in milliseconds). If `0`, the key will persist.
+//	ttl - The expiry time. If `0`, the key will persist.
 //	value - The serialized value to deserialize and assign to key.
 //
 // Command Response:
@@ -3648,7 +3655,7 @@ func (b *BaseBatch[T]) XGroupCreateWithOptions(key string, group string, id stri
 //	Return OK if successfully create a key with a value.
 //
 // [valkey.io]: https://valkey.io/commands/restore/
-func (b *BaseBatch[T]) Restore(key string, ttl int64, value string) *T {
+func (b *BaseBatch[T]) Restore(key string, ttl time.Duration, value string) *T {
 	return b.RestoreWithOptions(key, ttl, value, *options.NewRestoreOptions())
 }
 
@@ -3660,7 +3667,7 @@ func (b *BaseBatch[T]) Restore(key string, ttl int64, value string) *T {
 // Parameters:
 //
 //	key - The key to create.
-//	ttl - The expiry time (in milliseconds). If `0`, the key will persist.
+//	ttl - The expiry time. If `0`, the key will persist.
 //	value - The serialized value to deserialize and assign to key.
 //	restoreOptions - Set restore options with replace and absolute TTL modifiers, object idletime and frequency.
 //
@@ -3669,14 +3676,19 @@ func (b *BaseBatch[T]) Restore(key string, ttl int64, value string) *T {
 //	Return OK if successfully create a key with a value.
 //
 // [valkey.io]: https://valkey.io/commands/restore/
-func (b *BaseBatch[T]) RestoreWithOptions(key string, ttl int64, value string, restoreOptions options.RestoreOptions) *T {
+func (b *BaseBatch[T]) RestoreWithOptions(
+	key string,
+	ttl time.Duration,
+	value string,
+	restoreOptions options.RestoreOptions,
+) *T {
 	optionArgs, err := restoreOptions.ToArgs()
 	if err != nil {
 		return b.addError("RestoreWithOptions", err)
 	}
 	return b.addCmdAndTypeChecker(C.Restore, append([]string{
 		key,
-		utils.IntToString(ttl), value,
+		utils.IntToString(ttl.Milliseconds()), value,
 	}, optionArgs...), reflect.String, false)
 }
 
@@ -4262,17 +4274,17 @@ func (b *BaseBatch[T]) GetBit(key string, offset int64) *T {
 // Parameters:
 //
 //	numberOfReplicas - The number of replicas to reach.
-//	timeout - The timeout value specified in milliseconds. A value of `0` will block indefinitely.
+//	timeout - The timeout value. A value of `0` will block indefinitely.
 //
 // Command Response:
 //
 //	The number of replicas reached by all the writes performed in the context of the current connection.
 //
 // [valkey.io]: https://valkey.io/commands/wait/
-func (b *BaseBatch[T]) Wait(numberOfReplicas int64, timeout int64) *T {
+func (b *BaseBatch[T]) Wait(numberOfReplicas int64, timeout time.Duration) *T {
 	return b.addCmdAndTypeChecker(
 		C.Wait,
-		[]string{utils.IntToString(numberOfReplicas), utils.IntToString(timeout)},
+		[]string{utils.IntToString(numberOfReplicas), utils.IntToString(timeout.Milliseconds())},
 		reflect.Int64,
 		false,
 	)
@@ -4369,7 +4381,7 @@ func (b *BaseBatch[T]) BitCountWithOptions(key string, opts options.BitCountOpti
 //	the consumer.
 //
 // [valkey.io]: https://valkey.io/commands/xclaim/
-func (b *BaseBatch[T]) XClaim(key string, group string, consumer string, minIdleTime int64, ids []string) *T {
+func (b *BaseBatch[T]) XClaim(key string, group string, consumer string, minIdleTime time.Duration, ids []string) *T {
 	return b.XClaimWithOptions(key, group, consumer, minIdleTime, ids, *options.NewXClaimOptions())
 }
 
@@ -4395,11 +4407,11 @@ func (b *BaseBatch[T]) XClaimWithOptions(
 	key string,
 	group string,
 	consumer string,
-	minIdleTime int64,
+	minIdleTime time.Duration,
 	ids []string,
 	opts options.XClaimOptions,
 ) *T {
-	args := append([]string{key, group, consumer, utils.IntToString(minIdleTime)}, ids...)
+	args := append([]string{key, group, consumer, utils.IntToString(minIdleTime.Milliseconds())}, ids...)
 	optionArgs, err := opts.ToArgs()
 	if err != nil {
 		return b.addError("XClaimWithOptions", err)
@@ -4427,7 +4439,7 @@ func (b *BaseBatch[T]) XClaimWithOptions(
 //	the consumer.
 //
 // [valkey.io]: https://valkey.io/commands/xclaim/
-func (b *BaseBatch[T]) XClaimJustId(key string, group string, consumer string, minIdleTime int64, ids []string) *T {
+func (b *BaseBatch[T]) XClaimJustId(key string, group string, consumer string, minIdleTime time.Duration, ids []string) *T {
 	return b.XClaimJustIdWithOptions(key, group, consumer, minIdleTime, ids, *options.NewXClaimOptions())
 }
 
@@ -4454,11 +4466,11 @@ func (b *BaseBatch[T]) XClaimJustIdWithOptions(
 	key string,
 	group string,
 	consumer string,
-	minIdleTime int64,
+	minIdleTime time.Duration,
 	ids []string,
 	opts options.XClaimOptions,
 ) *T {
-	args := append([]string{key, group, consumer, utils.IntToString(minIdleTime)}, ids...)
+	args := append([]string{key, group, consumer, utils.IntToString(minIdleTime.Milliseconds())}, ids...)
 	optionArgs, err := opts.ToArgs()
 	if err != nil {
 		return b.addError("XClaimJustIdWithOptions", err)
@@ -5220,9 +5232,7 @@ func (b *BaseBatch[T]) ZMPop(keys []string, scoreFilter constants.ScoreFilter) *
 
 	// Check for potential length overflow.
 	if len(keys) > math.MaxInt-2 {
-		return b.addError("ZMPop", &errors.RequestError{
-			Msg: "Length overflow for the provided keys",
-		})
+		return b.addError("ZMPop", errors.New("length overflow for the provided keys"))
 	}
 
 	// args slice will have 2 more arguments with the keys provided.
@@ -5260,9 +5270,7 @@ func (b *BaseBatch[T]) ZMPopWithOptions(keys []string, scoreFilter constants.Sco
 
 	// Check for potential length overflow.
 	if len(keys) > math.MaxInt-4 {
-		return b.addError("ZMPopWithOptions", &errors.RequestError{
-			Msg: "Length overflow for the provided keys",
-		})
+		return b.addError("ZMPopWithOptions", errors.New("length overflow for the provided keys"))
 	}
 
 	// args slice will have 4 more arguments with the keys provided.